Consolidate color tests
